Extract code generation from lib.rs to modules
